+++ frontend/pages/gmm/index.tsx
@@ -5,8 +5,8 @@ import Head from "next/head";
 import { Col, Container, Row } from "react-bootstrap";
 import Navigator from "~/components/common/navigator";
 import Footer from "~/components/common/footer";
-import AddJobButton from "~/components/gmm/home/add-job-button";
-import GmmJobsList from "~/components/gmm/home/gmm-jobs-list";
+import AddJobButton from "~/components/gmm-home/add-job-button";
+import GmmJobsList from "~/components/gmm-home/gmm-jobs-list";
 import { useRouter } from "next/router";
 import { useEffect, useState } from "react";
 import { Alert, Button } from "react-bootstrap";
@@ -16,15 +16,15 @@ import { responseGetGMMJobsItems } from "~/services/route/gmm";
 import { ArrowClockwise } from "react-bootstrap-icons";
 import { useIsLoading } from "~/hooks/common";
 import LoadingPane from "~/components/common/loading-pane";
-import TimeDescription from "~/components/gmm/home/time-description";
-import { ActionButtons } from "~/components/gmm/home/action-buttons";
-import ParamsTable from "~/components/gmm/home/params-table";
-import { CurrentInfo } from "~/components/gmm/home/current-info";
-import { GmmNumComponentSelector } from "~/components/gmm/home/gmm-job-handler";
-import LatentGraph from "~/components/gmm/home/latent-graph";
-import BicGraph from "~/components/gmm/home/bic-graph";
+import TimeDescription from "~/components/gmm-home/time-description";
+import { ActionButtons } from "~/components/gmm-home/action-buttons";
+import ParamsTable from "~/components/gmm-home/params-table";
+import { CurrentInfo } from "~/components/gmm-home/current-info";
+import { GmmNumComponentSelector } from "~/components/gmm-home/gmm-job-handler";
+import LatentGraph from "~/components/gmm-home/latent-graph";
+import BicGraph from "~/components/gmm-home/bic-graph";
 import { Provider } from "react-redux";
-import { store } from "~/components/gmm/home/redux/store";
+import { store } from "~/components/gmm-home/redux/store";
 
 type JobItem = z.infer<typeof responseGetGMMJobsItems>;
