+++ frontend/pages/trainer/add.tsx
@@ -4,19 +4,19 @@ import Head from "next/head";
 import { Container } from "react-bootstrap";
 import Navigator from "~/components/common/navigator";
 import { Provider } from "react-redux";
-import { store } from "~/components/trainer/add/redux/store";
+import { store } from "~/components/trainer-add/redux/store";
 import "@inovua/reactdatagrid-community/index.css";
 import { useRouter } from "next/router";
 import Footer from "~/components/common/footer";
 import React from "react";
 import { Col, Row } from "react-bootstrap";
-import PreprocessingForms from "~/components/trainer/add/preprocessing-forms";
-import UploadFile from "~/components/trainer/add/upload-file";
-import SelexPagenation from "~/components/trainer/add/navigator-preprocess";
-import TrainPagenation from "~/components/trainer/add/navigator-train";
-import TrainParametersForms from "~/components/trainer/add/train-parameters-forms";
-import CountTable from "~/components/trainer/add/count-table";
-import SequenceTable from "~/components/trainer/add/sequence-table";
+import PreprocessingForms from "~/components/trainer-add/preprocessing-forms";
+import UploadFile from "~/components/trainer-add/upload-file";
+import SelexPagenation from "~/components/trainer-add/navigator-preprocess";
+import TrainPagenation from "~/components/trainer-add/navigator-train";
+import TrainParametersForms from "~/components/trainer-add/train-parameters-forms";
+import CountTable from "~/components/trainer-add/count-table";
+import SequenceTable from "~/components/trainer-add/sequence-table";
 
 const App: React.FC = () => {
   const router = useRouter();
